+++ sdk/python/feast/infra/compute_engines/local/feature_builder.py
@@ -1,5 +1,7 @@
 from typing import Union
 
+from feast.infra.common.materialization_job import MaterializationTask
+from feast.infra.common.retrieval_job import HistoricalRetrievalTask
 from feast.infra.compute_engines.dag.plan import ExecutionPlan
 from feast.infra.compute_engines.feature_builder import FeatureBuilder
 from feast.infra.compute_engines.local.backends.base import DataFrameBackend
@@ -13,8 +15,6 @@ from feast.infra.compute_engines.local.node import (
     LocalTransformationNode,
     LocalValidationNode,
 )
-from feast.infra.compute_engines.tasks import HistoricalRetrievalTask
-from feast.infra.materialization.batch_materialization_engine import MaterializationTask
 
 
 class LocalFeatureBuilder(FeatureBuilder):
