+++ sdk/python/feast/infra/compute_engines/spark/compute.py
@@ -1,13 +1,13 @@
-from feast.infra.compute_engines.base import ComputeEngine
-from feast.infra.compute_engines.spark.feature_builder import SparkFeatureBuilder
-from feast.infra.compute_engines.spark.job import SparkDAGRetrievalJob
-from feast.infra.compute_engines.spark.utils import get_or_create_new_spark_session
-from feast.infra.compute_engines.tasks import HistoricalRetrievalTask
-from feast.infra.materialization.batch_materialization_engine import (
+from feast.infra.common.materialization_job import (
     MaterializationJob,
     MaterializationJobStatus,
     MaterializationTask,
 )
+from feast.infra.common.retrieval_job import HistoricalRetrievalTask
+from feast.infra.compute_engines.base import ComputeEngine
+from feast.infra.compute_engines.spark.feature_builder import SparkFeatureBuilder
+from feast.infra.compute_engines.spark.job import SparkDAGRetrievalJob
+from feast.infra.compute_engines.spark.utils import get_or_create_new_spark_session
 from feast.infra.materialization.contrib.spark.spark_materialization_engine import (
     SparkMaterializationJob,
 )
