+++ sdk/python/feast/infra/materialization/batch_materialization_engine.py
@@ -1,68 +1,19 @@
-import enum
 from abc import ABC, abstractmethod
-from dataclasses import dataclass
-from datetime import datetime
-from typing import Callable, List, Optional, Sequence, Union
-
-from tqdm import tqdm
+from typing import List, Sequence, Union
 
 from feast.batch_feature_view import BatchFeatureView
 from feast.entity import Entity
 from feast.feature_view import FeatureView
+from feast.infra.common.materialization_job import (
+    MaterializationJob,
+    MaterializationTask,
+)
 from feast.infra.offline_stores.offline_store import OfflineStore
 from feast.infra.online_stores.online_store import OnlineStore
 from feast.infra.registry.base_registry import BaseRegistry
 from feast.on_demand_feature_view import OnDemandFeatureView
 from feast.repo_config import RepoConfig
 from feast.stream_feature_view import StreamFeatureView
-
-
-@dataclass
-class MaterializationTask:
-    """
-    A MaterializationTask represents a unit of data that needs to be materialized from an
-    offline store to an online store.
-    """
-
-    project: str
-    feature_view: Union[BatchFeatureView, StreamFeatureView, FeatureView]
-    start_time: datetime
-    end_time: datetime
-    tqdm_builder: Callable[[int], tqdm]
-
-
-class MaterializationJobStatus(enum.Enum):
-    WAITING = 1
-    RUNNING = 2
-    AVAILABLE = 3
-    ERROR = 4
-    CANCELLING = 5
-    CANCELLED = 6
-    SUCCEEDED = 7
-
-
-class MaterializationJob(ABC):
-    """
-    A MaterializationJob represents an ongoing or executed process that materializes data as per the
-    definition of a materialization task.
-    """
-
-    task: MaterializationTask
-
-    @abstractmethod
-    def status(self) -> MaterializationJobStatus: ...
-
-    @abstractmethod
-    def error(self) -> Optional[BaseException]: ...
-
-    @abstractmethod
-    def should_be_retried(self) -> bool: ...
-
-    @abstractmethod
-    def job_id(self) -> str: ...
-
-    @abstractmethod
-    def url(self) -> Optional[str]: ...
 
 
 class BatchMaterializationEngine(ABC):
