+++ sdk/python/feast/infra/materialization/kubernetes/k8s_materialization_job.py
@@ -2,9 +2,9 @@ from typing import Optional
 
 from kubernetes import client
 
+from feast.infra.common.materialization_job import MaterializationJobStatus
 from feast.infra.materialization.batch_materialization_engine import (
     MaterializationJob,
-    MaterializationJobStatus,
 )
