+++ sdk/python/feast/infra/materialization/local_engine.py
@@ -7,6 +7,11 @@ from tqdm import tqdm
 from feast.batch_feature_view import BatchFeatureView
 from feast.entity import Entity
 from feast.feature_view import FeatureView
+from feast.infra.common.materialization_job import (
+    MaterializationJob,
+    MaterializationJobStatus,
+    MaterializationTask,
+)
 from feast.infra.offline_stores.offline_store import OfflineStore
 from feast.infra.online_stores.online_store import OnlineStore
 from feast.infra.registry.base_registry import BaseRegistry
@@ -21,9 +26,6 @@ from feast.utils import (
 
 from .batch_materialization_engine import (
     BatchMaterializationEngine,
-    MaterializationJob,
-    MaterializationJobStatus,
-    MaterializationTask,
 )
 
 DEFAULT_BATCH_SIZE = 10_000
